// my-react-app/src/shared/types.tsx
type data = {
    initListFilms: [];
    currentListFilms: [];
    sortBy: string;
    genres: [];
    year: string;
    favorites: [];
    open: boolean;
};
type item = {
    adult: boolean;
    backdrop_path: string;
    genre_ids: [];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
};
export { type data, type item };